+++ src/eigen.cpp
@@ -59,7 +59,7 @@ int main (int argc, char **argv){
       if (mval.size() == 1 && mval.at(0) == 0){
         //If mode = 0 then print all eigenvalues
         mval.clear();
-        for (j=0; j< anin->getEigen().eigenvalues().rows(); j++){
+        for (j=0; j< static_cast<unsigned int>(anin->getEigen().eigenvalues().rows()); j++){
           mval.push_back(j+1);
         }
       }
@@ -77,7 +77,7 @@ int main (int argc, char **argv){
       if (mvec.size() == 1 && mvec.at(0) == 0){
         //If mode = 0 then print all eigenvectors
         mvec.clear();
-        for (j=0; j< anin->getEigen().eigenvectors().cols(); j++){
+        for (j=0; static_cast<unsigned int>(j< anin->getEigen().eigenvectors().cols()); j++){
           mvec.push_back(j+1);
         }
       }
